Create SMB DR Fix
This script enables the Windows Feature for SMB1 and 2, sets the reg entries to force enable, then disables the outdated or deprecated features, leaving the correct ones which are SMB v2. This was created in response to a patch by MS in 2021 which disabled SMB v1, but also removed machines abilities to communicate with the auth servers.